Scentiment/requirements.txt
openai
chainlit
loguru
python-dotenv